feat(whats-that-gerber): Use entire collection of filenames to determine types
whats-that-gerber is almost always used to determine the layer types of a collection of filenames
that represent an entire board. By changing wtg to accept / identify all files at once, it can make
a more informed decision based on which CAD pattern the collection appears to be following.

BREAKING CHANGE: Output of whats-that-gerber changed from a single string to an object keyed by
filename of layer type objects